Implement analyzing of instruction combinations.
Please don't judge code cleanliness, just wanted to get something
working out.
The parser was slightly modified to put pc+1 for unused registers, so
there are no false matches vs x0 or due to using random bits of an
immediate for registers. N_REG was increased accordingly to prevent
indexing problems.
Disclaimer: This is not the most memory friendly, although it allocates
lazily and uses mass allocation for the list nodes.
Displays how often a sequence of instructions with two specific
mnemonics, where rd of the first instruction is reused in the second
happens in the analyzed binary. Then it lists how often different
 combinations of register re-usage happen and the most used immediates
 for them.